requirements.txt
pdftotext==2.1.5
pysummarize==0.6.0
wordcloud==1.8.0
Pillow==8.0.0
